Subo funcionalidad de salida, faltaria agregar semaforos a todo lo que está, verificar bien como se alteran la cantidad de autos en el estacionamiento y el monto total. Ver si hay necesidad de hacer un kill de los procesos auto que no pueden entrar al estacionamiento. Se quitaron las clases entrada y salida, encapsulando todo el comportamiento de las mismas en la clase simulador entrada y simulador salida correspondientemente.

+++ Estructuras.h
@@ -5,7 +5,7 @@
 
 struct Parametros {
         int lugaresDisponibles;
-        int costoHora;
+        float costoHora;
         int tiempoSimulacion;
 };
